+++ packages/test-tabs/src/tests-generic.js
@@ -8,7 +8,7 @@ const iconSettingsSVG = "<svg width=\"24\" height=\"24\" viewBox=\"0 0 24.00 24.
 const arrowBackSVG = "<svg width=\"24\" height=\"24\" viewBox=\"0 0 24 24\"><path d=\"M20 11H7.83l5.59-5.59L12 4l-8 8 8 8 1.41-1.41L7.83 13H20v-2z\"/></svg>";
 const arrowForwardSVG = "<svg width=\"24\" height=\"24\" viewBox=\"0 0 24 24\"><path d=\"M12 4l-1.41 1.41L16.17 11H4v2h12.17l-5.58 5.59L12 20l8-8z\"/></svg>";
 
-export default ({ renderer: h, keys: k, Tabs }) => {
+export default ({ renderer: h, Tabs }) => {
 
   const iconHeart = h.trust(iconHeartSVG);
   const iconBell = h.trust(iconBellSVG);
@@ -156,12 +156,12 @@ export default ({ renderer: h, keys: k, Tabs }) => {
       }
     },
     {
-      name: "Option: selectedTab (1)",
+      name: "Option: selectedTabIndex (1)",
       component: Tabs,
       attrs: {
         tabs: threeButtons,
         autofit: true,
-        selectedTab: 1
+        selectedTabIndex: 1
       }
     },
     {
@@ -309,7 +309,7 @@ export default ({ renderer: h, keys: k, Tabs }) => {
       name: "Disabled tab",
       component: Tabs,
       attrs: {
-        selectedTab: 1,
+        selectedTabIndex: 1,
         autofit: true,
         content: [
           { label: "New", disabled: true },
@@ -371,7 +371,6 @@ export default ({ renderer: h, keys: k, Tabs }) => {
       exclude: true,
       component: OnChange
     },
-
     {
       name: "Option: content (RTL)",
       component: {